requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
amqp==5.1.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-vtk==0.0.9
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
diskcache==5.4.0
docutils==0.19
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi
fastjsonschema==2.16.2
Flask==2.2.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
japanize-kivy==0.1.1
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
multiprocess==0.70.14
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
plotly==5.11.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pygame==2.1.2
Pygments==2.13.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.4
pytz==2022.6
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
stack-data==0.6.2
starlette
tenacity==8.1.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tornado==6.2
traitlets==5.8.0
typing_extensions==4.2.0
uri-template==1.2.0
urllib3==1.26.12
uvicorn==0.17.6
vine==5.0.0
vtk==9.2.2
waitress==2.1.2
watchgod==0.8.2
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.3
Werkzeug==2.2.2
widgetsnbextension==4.0.5
windows-curses==2.3.0
wslink==1.9.1
XlsxWriter==3.0.3
yarl==1.8.1
